requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.9.0
asgiref==3.8.1
astor==0.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
constantly==23.10.4
crispy-bootstrap5==2025.6
cryptography==46.0.2
daphne==4.2.1
distro==1.9.0
dj-database-url==3.0.1
Django==4.2.21
django-admin-volt==1.0.17
django-axes==8.0.0
django-contrib-postgres==0.0.1
django-crispy-forms==2.4
django-dbbackup==4.2.1
django-debug-toolbar==4.4.6
django-extensions==4.1
django-filter==25.1
django-jazzmin==3.0.1
django-postgres-extra==2.0.8
django-redis==5.4.0
django-tables2==2.7.5
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
geographiclib==2.0
geopy==2.4.1
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.2
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
jiter==0.9.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.1
multidict==6.4.4
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.6.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.9.0
pyOpenSSL==25.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.27.1
rsa==4.9.1
service-identity==24.2.0
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
str2bool==1.1
tablib==3.9.0
tokenizers==0.21.1
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
whitenoise==6.7.0
yarl==1.20.0
zope.interface==8.0.1
